+++ mobile-app/cypress/e2e/functional/wallet/portfolio/send.spec.ts
@@ -424,253 +424,6 @@ context("Wallet - Send", { testIsolation: false }, () => {
   });
 });
 
-context("Wallet - Send - Max Values", { testIsolation: false }, () => {
-  let whale: WhaleApiClient;
-
-  before(() => {
-    cy.clearLocalStorage();
-    cy.clearCookies();
-  });
-
-  const addresses = [
-    "bcrt1q6np0fh47ykhznjhrtfvduh73cgjg32yac8t07d",
-    "bcrt1qyynghf6xv66c7zewd6aansn9j9hy3q2hsl7ms7",
-  ];
-  const prevBalances: { [key: string]: string } = {};
-  beforeEach(() => {
-    const network = localStorage.getItem("Development.NETWORK");
-    whale = new WhaleApiClient({
-      url:
-        network === "Playground"
-          ? "https://playground.jellyfishsdk.com"
-          : "http://localhost:19553",
-      network: "regtest",
-      version: "v0",
-    });
-    cy.createEmptyWallet(true);
-    cy.sendDFItoWallet().wait(4000);
-    cy.getByTestID("bottom_tab_portfolio").click();
-  });
-
-  addresses.forEach((address: string) => {
-    describe(`check for address ${address}`, () => {
-      it(`should be able to send to address ${address}`, () => {
-        cy.wrap(whale.address.getBalance(address)).then((response: any) => {
-          prevBalances[address] = response;
-        });
-        cy.getByTestID("bottom_tab_portfolio").click();
-        cy.getByTestID("portfolio_list").should("exist");
-        cy.getByTestID("dfi_total_balance_amount").contains("10.00000000");
-        cy.getByTestID("dfi_balance_card").should("exist");
-        cy.getByTestID("action_button_group").should("exist");
-        cy.getByTestID("send_balance_button").click().wait(3000);
-        cy.getByTestID("select_DFI").click().wait(3000);
-        cy.getByTestID("address_input").clear().type(address);
-        cy.getByTestID("MAX_amount_button").click();
-        cy.getByTestID("button_confirm_send_continue").should(
-          "not.have.attr",
-          "disabled",
-        );
-        cy.getByTestID("button_confirm_send_continue").click();
-        cy.getByTestID("confirm_title").contains("You are sending");
-        cy.getByTestID("text_send_amount").contains("9.90000000");
-        cy.getByTestID("text_amount").contains("9.90000000 DFI");
-        const usdValueWithThousandSep = Number(
-          new BigNumber(9.9).multipliedBy(10000).toFixed(2),
-        ).toLocaleString(undefined, {
-          minimumFractionDigits: 2,
-          maximumFractionDigits: 2,
-        });
-        cy.getByTestID("text_amount_rhsUsdAmount").should(
-          "have.text",
-          `$${usdValueWithThousandSep}`,
-        );
-        cy.getByTestID("button_confirm_send").click().wait(3000);
-        cy.closeOceanInterface();
-        cy.getByTestID("bottom_tab_portfolio").click();
-        cy.getByTestID("dfi_total_balance_amount").contains("0.09");
-      });
-
-      it(`should check if exist on other side ${address}`, () => {
-        cy.wrap(whale.address.getBalance(address)).then((response) => {
-          const updatedBalance = new BigNumber(response as string).minus(
-            prevBalances[address],
-          );
-          expect(updatedBalance.toString()).contains("9.9");
-        });
-      });
-    });
-  });
-});
-
-context("Wallet - Send - with Conversion", { testIsolation: false }, () => {
-  let whale: WhaleApiClient;
-  before(() => {
-    cy.clearLocalStorage();
-    cy.clearCookies();
-  });
-
-  const addresses = [
-    "bcrt1qh5callw3zuxtnks96ejtekwue04jtnm84f04fn",
-    "bcrt1q6ey8k3w0ll3cn5sg628nxthymd3une2my04j4n",
-  ];
-  const prevBalances: { [key: string]: string } = {};
-
-  beforeEach(() => {
-    const network = localStorage.getItem("Development.NETWORK");
-    whale = new WhaleApiClient({
-      url:
-        network === "Playground"
-          ? "https://playground.jellyfishsdk.com"
-          : "http://localhost:19553",
-      network: "regtest",
-      version: "v0",
-    });
-  });
-
-  addresses.forEach((address) => {
-    describe(`check for address ${address}`, () => {
-      it(`should be able to send to address ${address}`, () => {
-        cy.createEmptyWallet(true);
-        cy.sendDFItoWallet().sendDFITokentoWallet();
-        cy.wait(6000);
-        cy.getByTestID("bottom_tab_portfolio").click();
-        cy.wrap(whale.address.getBalance(address)).then((response: any) => {
-          prevBalances[address] = response;
-        });
-        cy.getByTestID("bottom_tab_portfolio").click();
-        cy.getByTestID("portfolio_list").should("exist");
-        cy.getByTestID("dfi_balance_card").should("exist");
-        cy.getByTestID("action_button_group").should("exist");
-        cy.getByTestID("send_balance_button").click().wait(3000);
-        cy.getByTestID("select_DFI").click().wait(3000);
-        cy.getByTestID("address_input").clear().type(address);
-        cy.getByTestID("amount_input").type("12");
-        cy.getByTestID("transaction_details_info_text").should(
-          "contain",
-          "By continuing, the required amount of DFI will be converted",
-        );
-
-        cy.getByTestID("button_confirm_send_continue").click();
-        cy.getByTestID("txn_authorization_title").contains(
-          `Convert ${new BigNumber("2.1").toFixed(8)} DFI to UTXO`,
-        );
-        cy.closeOceanInterface().wait(3000);
-
-        cy.getByTestID("amount_to_convert_label").should(
-          "have.text",
-          "Amount to convert",
-        );
-        cy.getByTestID("amount_to_convert_value").should(
-          "contain",
-          "2.10000000 DFI",
-        );
-        cy.getByTestID("conversion_status").should("have.text", "Converted");
-        cy.getByTestID("transaction_fee_label").should(
-          "have.text",
-          "Transaction fee",
-        );
-        cy.getByTestID("transaction_fee_value").should("exist");
-
-        cy.getByTestID("text_amount").should("have.text", "12.00000000 DFI");
-        const usdValueWithThousandSep = Number(
-          new BigNumber(12).multipliedBy(10000).toFixed(2),
-        ).toLocaleString(undefined, {
-          minimumFractionDigits: 2,
-          maximumFractionDigits: 2,
-        });
-        cy.getByTestID("text_amount_rhsUsdAmount").should(
-          "have.text",
-          `$${usdValueWithThousandSep}`,
-        );
-
-        cy.getByTestID("text_send_amount").should("contain", "12.00000000");
-        cy.getByTestID("button_confirm_send").click().wait(3000);
-        cy.closeOceanInterface();
-      });
-
-      it(`should check if exist on other side ${address}`, () => {
-        cy.wrap(whale.address.getBalance(address)).then((response) => {
-          const updatedBalance = new BigNumber(response as string).minus(
-            prevBalances[address],
-          );
-          expect(updatedBalance.toString()).contains("12");
-        });
-      });
-    });
-  });
-});
-
-context("Wallet - Send - Switch token", { testIsolation: false }, () => {
-  before(() => {
-    cy.clearLocalStorage();
-    cy.clearCookies();
-    cy.createEmptyWallet(true);
-    cy.sendDFItoWallet()
-      .sendDFITokentoWallet()
-      .sendTokenToWallet(["BTC", "ETH"])
-      .wait(6000);
-    cy.getByTestID("bottom_tab_portfolio").click();
-  });
-
-  it("should be able to select token", () => {
-    cy.getByTestID("send_balance_button").click();
-    cy.getByTestID("select_DFI_value").should("have.text", "19.90000000");
-    cy.getByTestID("select_dBTC_value").should("have.text", "10.00000000");
-    cy.getByTestID("select_dETH_value").should("have.text", "10.00000000");
-    cy.wait(3000); // timeout to allow max. one block-cycle of re-render
-    cy.getByTestID("select_DFI").click();
-    cy.getByTestID("button_confirm_send_continue").should(
-      "have.attr",
-      "aria-disabled",
-    );
-    cy.getByTestID("max_value").should("have.text", "19.90000000");
-    cy.getByTestID("max_value_display_symbol").contains("DFI");
-  });
-
-  it("should be able to switch token", () => {
-    cy.getByTestID("select_token_input").click();
-    cy.wait(3000); // timeout to allow max. one block-cycle of re-render
-    cy.getByTestID("select_dBTC").click();
-    cy.getByTestID("max_value").should("have.text", "10.00000000");
-    cy.getByTestID("max_value_display_symbol").contains("dBTC");
-    cy.getByTestID("button_confirm_send_continue").should(
-      "have.attr",
-      "aria-disabled",
-    );
-  });
-
-  it("should be able to pre-select token", () => {
-    cy.getByTestID("bottom_tab_portfolio").click();
-    cy.getByTestID("portfolio_list").should("exist");
-    cy.getByTestID("portfolio_row_2").should("exist");
-    cy.getByTestID("portfolio_row_2_amount").contains(10).click();
-    cy.getByTestID("send_button").click();
-    cy.getByTestID("max_value").should("have.text", "10.00000000");
-    cy.getByTestID("max_value_display_symbol").contains("dETH");
-  });
-
-  it("should be able to enable/disable token selection", () => {
-    cy.createEmptyWallet(true);
-    cy.wait(3000);
-    // No token
-    cy.getByTestID("bottom_tab_portfolio").should("exist");
-    cy.getByTestID("action_button_group").should("exist");
-    cy.getByTestID("send_balance_button").click().wait(3000);
-    cy.getByTestID("no_asset_text").should("exist");
-    cy.getByTestID("no_asset_sub_text").should("exist");
-
-    // With DFI
-    cy.getByTestID("bottom_tab_portfolio").click();
-    cy.sendDFITokentoWallet().wait(3000);
-    cy.getByTestID("action_button_group").should("exist");
-    cy.getByTestID("send_balance_button").click().wait(3000);
-    cy.getByTestID("select_DFI").click().wait(3000);
-    cy.getByTestID("max_value").should("have.text", "9.90000000");
-    cy.getByTestID("max_value_display_symbol").contains("DFI");
-  });
-});
-
 context("Wallet - Send - Address book", { testIsolation: false }, () => {
   function populateAddressBook(): void {
     cy.clearLocalStorage();
